refactoring of analysis file into directory
